pyproject.toml
[project]
name = "agentic-fetch"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "html2text>=2024",
    "beautifulsoup4>=4.12",
    "zendriver>=0.4",
    "pydantic-settings>=2.5",
    "pyyaml>=6.0",
    "markdownify>=0.13",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-asyncio>=0.24", "respx>=0.21", "ruff>=0.6"]

[project.scripts]
agentic-search = "agentic_fetch.cli.search:main"
agentic-fetch = "agentic_fetch.cli.fetch:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_fetch"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]
pythonpath = ["src"]
